Add single HabitTracker component for DRYness
