+++ src/routes/MentalHealthTracker.jsx
@@ -1,10 +1,19 @@
-const MentalHealthTracker = () => {
-  return (
-    <div className="p-4">
-      <h1 className="text-2xl font-bold mb-4">Mental Health Tracker</h1>
-      {/* Add your mental health tracking components here */}
-      <p>Track your mental health activities here.</p>
-    </div>
-  );
+import HabitTracker from "../components/HabitTracker";
+import mentalImage from "../assets/codingImage.jpg"; // Use your own image
+
+const mentalConfig = {
+  title: "Your Mental Health Hive",
+  imageSrc: mentalImage,
+  entryLabel: "How many minutes did you meditate or reflect?",
+  placeholder: "e.g. 15",
+  unit: "min",
+  mosaicGridSize: 4,
+  clearWarning:
+    "Are you sure you want to clear all your mental health data? This cannot be undone.",
 };
+
+const MentalHealthTracker = ({ entries, setEntries }) => (
+  <HabitTracker entries={entries} setEntries={setEntries} {...mentalConfig} />
+);
+
 export default MentalHealthTracker;
